+++ CHANGELOG.rst
@@ -2,6 +2,8 @@ Unreleased
 ----------
 * BREAKING CHANGE: the output of ``insturments.search`` function has changed from DataFrame to a list of dicts. Use ``old_search`` if you want to continue using the old format.
 * BREAKING CHANGE: renamed the ``c_sec_val`` parameter of ``Instrument.related_companies`` to ``cs``.
+* BREAKING CHANGE: The DataFrame result of ``Instrument.daily_closing_price()`` now has datetime index instead of ``dEven`` and ``hEven`` columns.
+* BREAKING CHANGE: ``Instrument.codal()`` now returns a list of dicts instead a dataframe.
 * Added new async property: ``Instrument.cs``
 * Added new async function: ``general.market_overview()``
 * Added new async method: ``Instrument.identity()``